feat: Strategy/Health component
