+++ components/StrategiesToBorrowFrom/StrategiesToBorrowFrom.tsx
@@ -1,4 +1,5 @@
 import { Fieldset } from 'components/Fieldset';
+import { Health } from 'components/Strategy/Health';
 import { ethers } from 'ethers';
 import { useConfig } from 'hooks/useConfig';
 import { LendingStrategy } from 'lib/LendingStrategy';
@@ -16,33 +17,6 @@ import {
   RateTooltip,
   TokenTooltip,
 } from './Tooltips';
-
-const percentDiff = (a: number, b: number): number =>
-  Math.abs(((b - a) / a) * 100);
-
-const diffToPosition = (diff: number, index: number, val: number) => {
-  if (val <= index) {
-    if (diff < 5) {
-      return 0;
-    } else if (diff < 10) {
-      return 1;
-    } else if (diff < 15) {
-      return 2;
-    } else {
-      return 3;
-    }
-  } else {
-    if (diff < 5) {
-      return 5;
-    } else if (diff < 10) {
-      return 6;
-    } else if (diff < 15) {
-      return 7;
-    } else {
-      return 8;
-    }
-  }
-};
 
 export type StrategiesToBorrowFromProps = {
   strategies: LendingStrategy[];
@@ -123,7 +97,6 @@ export default function StrategiesToBorrowFrom({
           <tbody>
             {strategies.map((strategy, i) => {
               const pricesDataForStrategy = pricesData[strategy.id];
-              const index = pricesDataForStrategy.index;
               const targetYearlyGrowth = pricesDataForStrategy.indexDPR * 365;
 
               const mark = parseFloat(
@@ -137,16 +110,6 @@ export default function StrategiesToBorrowFrom({
                 ],
               );
               const markOverNorm = mark / norm;
-
-              const indexVersusMark = percentDiff(index, mark);
-              const indexVersusNorm = percentDiff(index, norm);
-
-              let markPosition = diffToPosition(indexVersusMark, index, mark);
-              let normPosition = diffToPosition(indexVersusNorm, index, norm);
-
-              if (markPosition === normPosition) {
-                markPosition++;
-              }
 
               const fakeNFTValue = 300000;
               const debtTokenMarketCap =
@@ -213,17 +176,7 @@ export default function StrategiesToBorrowFrom({
                   </td>
                   <td className={styles.rate}>
                     <TooltipReference {...rateTooltip}>
-                      {['-', '-', '-', '-', '|', '-', '-', '-', '-'].map(
-                        (char, i) => (
-                          <>
-                            {markPosition === i && <>R</>}
-                            {normPosition === i && <>C</>}
-                            {markPosition !== i && normPosition !== i && (
-                              <>{char}</>
-                            )}
-                          </>
-                        ),
-                      )}
+                      <Health pricesData={pricesDataForStrategy} />
                     </TooltipReference>
                     <RateTooltip
                       strategy={strategy}
