+++ components/Strategy/Health/index.ts
@@ -1,0 +1,1 @@
+export { Health } from './Health';
